requirements.txt
anthropic>=0.15.0
flask>=2.2.3
flask-cors>=4.0.0
pandas>=2.0.0
plotly>=5.13.1
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
langchain>=0.1.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.20 